#4 퍼블리싱 - 이외 스타일 수정
