+++ package.json
@@ -11,9 +11,11 @@
     "test": "vitest --reporter verbose"
   },
   "dependencies": {
+    "prettier": "^3.4.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^7.1.1",
+    "vite-plugin-svgr": "^4.3.0",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
