+++ src/App.tsx
@@ -1,24 +1,5 @@
-import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import Home from "./pages/Home";
-import Game from "./pages/Game";
-import RootLayout from "./layouts/RootLayout";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <RootLayout />,
-    children: [
-      {
-        path: "/",
-        element: <Home />,
-      },
-      {
-        path: "/game",
-        element: <Game />,
-      },
-    ],
-  },
-]);
+import { RouterProvider } from "react-router-dom";
+import router from "./Router";
 
 const App = () => {
   return <RouterProvider router={router} />;
