--- src/pages/Game.tsx
@@ -1,4 +1,0 @@
-const Game = () => {
-  return <div>Game</div>;
-};
-export default Game;
