--- src/pages/Home.tsx
@@ -1,16 +1,0 @@
-const Home = () => {
-  return (
-    <div className="container">
-      <div className="blue-marble-container">
-        <header>
-          <div>Logo</div>
-          <h1 className="game-title">BlueMarble Mate</h1>
-        </header>
-        <main className="p-4">
-          <p className="game-subtitle">블루마블을 저와함께 편하게 즐겨보세요</p>
-        </main>
-      </div>
-    </div>
-  );
-};
-export default Home;
