+++ src/styles/base/_reset.scss
@@ -48,3 +48,10 @@ table {
 	border-collapse: collapse;
 	border-spacing: 0;
 }
+button {
+  appearance: none;
+  background: none;
+  border: none;
+  padding: 0;
+  margin: 0;
+}
