+++ src/styles/utilities/_animations.scss
@@ -1,0 +1,17 @@
+@keyframes rolling {
+  0% {
+    transform: rotateX(0deg) rotateY(0deg) rotateZ(0deg);
+  }
+  25% {
+    transform: rotateX(90deg) rotateY(45deg) rotateZ(45deg);
+  }
+  50% {
+    transform: rotateX(180deg) rotateY(90deg) rotateZ(90deg);
+  }
+  75% {
+    transform: rotateX(270deg) rotateY(135deg) rotateZ(135deg);
+  }
+  100% {
+    transform: rotateX(360deg) rotateY(180deg) rotateZ(180deg);
+  }
+}
