+++ src/types/svg.d.ts
@@ -1,0 +1,5 @@
+declare module "*.svg?react" {
+  import React = require("react");
+  const SVG: React.FC<React.SVGProps<SVGSVGElement>>;
+  export default SVG;
+}
