+++ vite.config.ts
@@ -1,10 +1,11 @@
 import react from "@vitejs/plugin-react";
+import svgr from "vite-plugin-svgr";
 
 /// <reference types="vitest/config" />
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  plugins: [react()],
+  plugins: [react(), svgr()],
   test: {
     environment: "jsdom",
     globals: true,
